build success; refactored interfaces into classes and adjusted PriorityQueue
