+++ packages/apis/uniswapv2/src/query/pair.ts
@@ -1,6 +1,6 @@
 import { Ethereum_Query } from "./w3/imported";
 import { fetchTokenData } from "./fetch";
-import { tokenEquals, tokenSortsBefore } from "./token";
+import { tokenSortsBefore } from "./token";
 import { factoryAddress, minimumLiquidity } from "./index";
 import {
   Input_pairAddress,
@@ -18,11 +18,7 @@ import {
 } from "./w3";
 
 import { BigInt } from "as-bigint";
-
-interface ProcessedPair {
-  amount: TokenAmount;
-  nextPair: Pair;
-}
+import { ProcessedPair } from "../utils/ProcessedPair";
 
 // TODO: this can be calculated off-chain with keccack256
 // returns address of pair liquidity token contract
@@ -61,28 +57,28 @@ export function pairReserves(input: Input_pairReserves): TokenAmount[] {
 export function pairOutputAmount(input: Input_pairOutputAmount): TokenAmount {
   const pair: Pair = input.pair;
   const tradeTokenAmount: TokenAmount = input.inputAmount;
-  return pairOutput(pair, tradeTokenAmount).amount;
+  return ProcessedPair.pairOutputFromExactInput(pair, tradeTokenAmount).amount;
 }
 
 // Pricing function for exact input amounts. Returns next pair state, based on current reserves, if the trade were executed.
 export function pairOutputNextPair(input: Input_pairOutputNextPair): Pair {
   const pair: Pair = input.pair;
   const tradeTokenAmount: TokenAmount = input.inputAmount;
-  return pairOutput(pair, tradeTokenAmount).nextPair;
+  return ProcessedPair.pairOutputFromExactInput(pair, tradeTokenAmount).nextPair;
 }
 
 // Pricing function for exact output amounts. Returns minimum input amount, based on current reserves, if the trade were executed.
 export function pairInputAmount(input: Input_pairInputAmount): TokenAmount {
   const pair: Pair = input.pair;
   const tradeTokenAmount: TokenAmount = input.outputAmount;
-  return pairInput(pair, tradeTokenAmount).amount;
+  return ProcessedPair.pairInputForExactOutput(pair, tradeTokenAmount).amount;
 }
 
 // Pricing function for exact output amounts. Returns next pair state, based on current reserves, if the trade were executed.
 export function pairInputNextPair(input: Input_pairInputNextPair): Pair {
   const pair: Pair = input.pair;
   const tradeTokenAmount: TokenAmount = input.outputAmount;
-  return pairInput(pair, tradeTokenAmount).nextPair;
+  return ProcessedPair.pairInputForExactOutput(pair, tradeTokenAmount).nextPair;
 }
 
 /*
@@ -177,118 +173,4 @@ export function pairLiquidityValue(
     { token: pair.tokenAmount0.token, amount: token0Value.toString() },
     { token: pair.tokenAmount1.token, amount: token1Value.toString() },
   ];
-}
-
-function pairOutput(pair: Pair, tradeTokenAmount: TokenAmount): ProcessedPair {
-  const tradeAmount = BigInt.fromString(tradeTokenAmount.amount);
-  if (tradeAmount.eq(BigInt.ZERO)) {
-    throw new RangeError(
-      "Insufficient input amount: Input amount must be greater than zero"
-    );
-  }
-  if (
-    BigInt.fromString(pair.tokenAmount0.amount).eq(BigInt.ZERO) ||
-    BigInt.fromString(pair.tokenAmount1.amount).eq(BigInt.ZERO)
-  ) {
-    throw new RangeError(
-      "Insufficient liquidity: Pair reserves must be greater than zero"
-    );
-  }
-  let inTokenAmount: TokenAmount;
-  let outTokenAmount: TokenAmount;
-  if (
-    tokenEquals({
-      token: pair.tokenAmount0.token,
-      other: tradeTokenAmount.token,
-    })
-  ) {
-    inTokenAmount = pair.tokenAmount0;
-    outTokenAmount = pair.tokenAmount1;
-  } else {
-    inTokenAmount = pair.tokenAmount1;
-    outTokenAmount = pair.tokenAmount0;
-  }
-  const biInTokenAmt: BigInt = BigInt.fromString(inTokenAmount.amount);
-  const biOutTokenAmt: BigInt = BigInt.fromString(outTokenAmount.amount);
-  const amountInWithFee: BigInt = tradeAmount.mul(BigInt.fromString("997"));
-  const numerator: BigInt = amountInWithFee.mul(biOutTokenAmt);
-  const denominator: BigInt = biInTokenAmt
-    .mul(BigInt.fromString("1000"))
-    .add(amountInWithFee);
-  const output = numerator.div(denominator);
-  return {
-    amount: {
-      token: outTokenAmount.token,
-      amount: output.toString(),
-    },
-    nextPair: {
-      tokenAmount0: {
-        token: inTokenAmount.token,
-        amount: biInTokenAmt.add(tradeAmount).toString(),
-      },
-      tokenAmount1: {
-        token: outTokenAmount.token,
-        amount: biOutTokenAmt.sub(output).toString(),
-      },
-    },
-  };
-}
-
-function pairInput(pair: Pair, tradeTokenAmount: TokenAmount): ProcessedPair {
-  const tradeAmount = BigInt.fromString(tradeTokenAmount.amount);
-  if (tradeAmount.eq(BigInt.ZERO)) {
-    throw new RangeError(
-      "Insufficient output amount: Output amount must be greater than zero"
-    );
-  }
-  if (
-    BigInt.fromString(pair.tokenAmount0.amount) == BigInt.ZERO ||
-    BigInt.fromString(pair.tokenAmount1.amount) == BigInt.ZERO
-  ) {
-    throw new RangeError(
-      "Insufficient liquidity: Pair reserves must be greater than zero"
-    );
-  }
-  let inTokenAmount: TokenAmount;
-  let outTokenAmount: TokenAmount;
-  if (
-    tokenEquals({
-      token: pair.tokenAmount0.token,
-      other: tradeTokenAmount.token,
-    })
-  ) {
-    outTokenAmount = pair.tokenAmount0;
-    inTokenAmount = pair.tokenAmount1;
-  } else {
-    outTokenAmount = pair.tokenAmount1;
-    inTokenAmount = pair.tokenAmount0;
-  }
-  const biInTokenAmt = BigInt.fromString(inTokenAmount.amount);
-  const biOutTokenAmt = BigInt.fromString(outTokenAmount.amount);
-  const numerator: BigInt = biInTokenAmt
-    .mul(tradeAmount)
-    .mul(BigInt.fromString("1000"));
-  const denominator: BigInt = biOutTokenAmt
-    .sub(tradeAmount)
-    .mul(BigInt.fromString("997"));
-  const input: BigInt = numerator
-    .div(denominator)
-    .add(BigInt.fromString("1"));
-
-  return {
-    amount: {
-      token: inTokenAmount.token,
-      amount: input.toString(),
-    },
-    nextPair: {
-      tokenAmount0: {
-        token: inTokenAmount.token,
-        amount: biInTokenAmt.add(input).toString(),
-      },
-      tokenAmount1: {
-        token: outTokenAmount.token,
-        amount: biOutTokenAmt.sub(tradeAmount).toString(),
-      },
-    },
-  };
 }
